Extract all label validation-related functions and move them into lgr_validator app
This app already contains a function called from lgr_editor
(validate_label_csv, which returns validation output as a CSV document).
We might as well place all validation related code there, and remove
duplication of validate_label functions.
