+++ src/lgr_validator/urls.py
@@ -13,7 +13,10 @@ urlpatterns = patterns('',
     url(r'^eval/{}/csv/'.format(LGR_SLUG_FORMAT_WITH_OPT_SET),
         views.validate_label_csv,
         name='lgr_validate_csv'),
-    url(r'^eval/{}/'.format(LGR_SLUG_FORMAT_WITH_OPT_SET),
+    url(r'^eval/{}/validate/$'.format(LGR_SLUG_FORMAT_WITH_OPT_SET),
         views.validate_label,
-        name='lgr_validate'),
+        name='lgr_validate_label'),
+    url(r'^eval/{}/validate-nf/$'.format(LGR_SLUG_FORMAT_WITH_OPT_SET),
+        views.validate_label_noframe,
+        name='lgr_validate_label_noframe'),
 )
